+++ backend/internal/web/httpRouter.go
@@ -8,6 +8,13 @@ import (
 	"github.com/gtsteffaniak/filebrowser/backend/pkg/settings"
 )
 
+const (
+	time60s = 60 * time.Second
+	time30s = 30 * time.Second
+	time10s = 10 * time.Second
+	time5s  = 5 * time.Second
+)
+
 // configureHTTPRouter registers all API, public, and static routes.
 func configureHTTPRouter(router, api, publicRoutes, publicApi *http.ServeMux) {
 	// health routes
@@ -58,8 +65,8 @@ func configureHTTPRouter(router, api, publicRoutes, publicApi *http.ServeMux) {
 	api.HandleFunc("POST /resources/archive", withUser(archiveCreateHandler))
 	api.HandleFunc("POST /resources/unarchive", withUser(unarchiveHandler))
 	api.HandleFunc("GET /resources/download", withUser(downloadHandler))
-	api.HandleFunc("GET /resources/view", withTimeout(mediaHandlerTimeout, withUserHelper(viewHandler)))
-	api.HandleFunc("GET /resources/preview", withTimeout(30*time.Second, withUserHelper(previewHandler)))
+	api.HandleFunc("GET /resources/view", withTimeout(time60s, withUserHelper(viewHandler)))
+	api.HandleFunc("GET /resources/preview", withTimeout(time30s, withUserHelper(previewHandler)))
 	api.HandleFunc("POST /resources/pause", withUser(resourcePauseHandler))
 	publicApi.HandleFunc("GET /resources", withHashFile(publicGetResourceHandler))
 	publicApi.HandleFunc("GET /resources/items", withHashFile(publicItemsGetHandler))
@@ -69,8 +76,8 @@ func configureHTTPRouter(router, api, publicRoutes, publicApi *http.ServeMux) {
 	publicApi.HandleFunc("DELETE /resources/bulk", withHashFile(publicBulkDeleteHandler))
 	publicApi.HandleFunc("PATCH /resources", withHashFile(publicPatchHandler))
 	publicApi.HandleFunc("GET /resources/download", withHashFile(publicDownloadHandler))
-	publicApi.HandleFunc("GET /resources/view", withTimeout(mediaHandlerTimeout, withHashFileHelper(PublicViewHandler)))
-	publicApi.HandleFunc("GET /resources/preview", withTimeout(30*time.Second, withHashFileHelper(publicPreviewHandler)))
+	publicApi.HandleFunc("GET /resources/view", withTimeout(time60s, withHashFileHelper(PublicViewHandler)))
+	publicApi.HandleFunc("GET /resources/preview", withTimeout(time30s, withHashFileHelper(publicPreviewHandler)))
 	publicApi.HandleFunc("POST /resources/pause", withHashFile(PublicPauseHandler))
 	// Legacy routes (backwards compatibility)
 	api.HandleFunc("GET /raw", withUser(downloadHandler))
@@ -105,9 +112,10 @@ func configureHTTPRouter(router, api, publicRoutes, publicApi *http.ServeMux) {
 	// ========================================
 	api.HandleFunc("GET /settings", withAdmin(settingsGetHandler))
 	api.HandleFunc("GET /settings/config", withAdmin(settingsConfigHandler))
-	api.HandleFunc("GET /settings/analytics", withAdmin(settingsAnalyticsGetHandler))
-	api.HandleFunc("PATCH /settings/analytics", withAdmin(settingsAnalyticsPatchHandler))
-	api.HandleFunc("GET /settings/analytics/preview", withAdmin(settingsAnalyticsPreviewHandler))
+	api.HandleFunc("GET /settings/analytics", withTimeout(time5s, withAdminHelper(settingsAnalyticsGetHandler)))
+	api.HandleFunc("PUT /settings/analytics", withTimeout(time5s, withAdminHelper(settingsAnalyticsPatchHandler)))
+	api.HandleFunc("PATCH /settings/analytics", withTimeout(time5s, withAdminHelper(settingsAnalyticsPatchHandler)))
+	api.HandleFunc("GET /settings/analytics/preview", withTimeout(time30s, withAdminHelper(settingsAnalyticsPreviewHandler)))
 	api.HandleFunc("GET /settings/sources", withUser(getSourceInfoHandler))
 
 	// ========================================
@@ -124,13 +132,13 @@ func configureHTTPRouter(router, api, publicRoutes, publicApi *http.ServeMux) {
 	// ========================================
 	// Media Routes - /api/media/ (with public routes)
 	// ========================================
-	api.HandleFunc("GET /media/subtitles", withTimeout(mediaHandlerTimeout, withUserHelper(subtitlesHandler)))
-	api.HandleFunc("GET /media/metadata", withTimeout(mediaHandlerTimeout, withUserHelper(metadataHandler)))
-	api.HandleFunc("GET /media/lyrics", withTimeout(mediaHandlerTimeout, withUserHelper(lyricsHandler)))
-	api.HandleFunc("GET /media/stream", withTimeout(mediaHandlerTimeout, withUserHelper(streamHandler)))
-	publicApi.HandleFunc("GET /media/metadata", withTimeout(mediaHandlerTimeout, withHashFileHelper(publicMetadataHandler)))
-	publicApi.HandleFunc("GET /media/lyrics", withTimeout(mediaHandlerTimeout, withHashFileHelper(publicLyricsHandler)))
-	publicApi.HandleFunc("GET /media/stream", withTimeout(mediaHandlerTimeout, withHashFileHelper(publicStreamHandler)))
+	api.HandleFunc("GET /media/subtitles", withTimeout(time60s, withUserHelper(subtitlesHandler)))
+	api.HandleFunc("GET /media/metadata", withTimeout(time60s, withUserHelper(metadataHandler)))
+	api.HandleFunc("GET /media/lyrics", withTimeout(time60s, withUserHelper(lyricsHandler)))
+	api.HandleFunc("GET /media/stream", withTimeout(time60s, withUserHelper(streamHandler)))
+	publicApi.HandleFunc("GET /media/metadata", withTimeout(time60s, withHashFileHelper(publicMetadataHandler)))
+	publicApi.HandleFunc("GET /media/lyrics", withTimeout(time60s, withHashFileHelper(publicLyricsHandler)))
+	publicApi.HandleFunc("GET /media/stream", withTimeout(time60s, withHashFileHelper(publicStreamHandler)))
 
 	// ========================================
 	// OnlyOffice Routes - /api/office/ (with public routes)
