--- backend/internal/web/public.go
@@ -1,510 +1,0 @@
-package web
-
-import (
-	"encoding/json"
-	"fmt"
-	"net/http"
-	"strings"
-
-	"github.com/gtsteffaniak/filebrowser/backend/internal/adapters/fs/files"
-	"github.com/gtsteffaniak/filebrowser/backend/internal/database/users"
-	"github.com/gtsteffaniak/filebrowser/backend/internal/errors"
-	"github.com/gtsteffaniak/filebrowser/backend/internal/preview"
-	"github.com/gtsteffaniak/filebrowser/backend/internal/state"
-	"github.com/gtsteffaniak/filebrowser/backend/internal/utils"
-	"github.com/gtsteffaniak/filebrowser/backend/pkg/settings"
-	"github.com/gtsteffaniak/go-logger/logger"
-)
-
-// publicDownloadHandler serves the raw content of a file, multiple files, or directory via a public share.
-// @Summary Download files from a public share
-// @Description Downloads raw content from a public share. Supports single files, multiple files, or directories as archives. Enforces download limits (global or per-user) and blocks anonymous users when per-user limits are enabled.
-// @Description
-// @Description **Multiple Files:**
-// @Description - Use repeated query parameters: `?file=file1.txt&file=file2.txt&file=file3.txt`
-// @Description - This supports filenames containing commas and special characters
-// @Tags Resources
-// @Accept json
-// @Produce octet-stream
-// @Param hash query string true "Share hash for authentication"
-// @Param file query []string true "File path (can be repeated for multiple files)"
-// @Param inline query bool false "If true, sets 'Content-Disposition' to 'inline'. Otherwise, defaults to 'attachment'."
-// @Param algo query string false "Compression algorithm for archiving multiple files or directories. Options: 'zip' and 'tar.gz'. Default is 'zip'."
-// @Success 200 {file} file "Raw file or directory content, or archive for multiple files"
-// @Failure 400 {object} map[string]string "Invalid request path or encoding"
-// @Failure 403 {object} map[string]string "Download limit reached, anonymous access blocked, or share unavailable"
-// @Failure 404 {object} map[string]string "Share not found or file not found"
-// @Failure 500 {object} map[string]string "Internal server error"
-// @Failure 501 {object} map[string]string "Downloads disabled for upload shares"
-// @Router /public/api/resources/download [get]
-func publicDownloadHandler(w http.ResponseWriter, r *http.Request, d *requestContext) (int, error) {
-	if d.Share.ShareType == "upload" {
-		return http.StatusNotImplemented, fmt.Errorf("downloads are disabled for upload shares")
-	}
-
-	// Check DisableDownload permission for normal shares
-	if d.Share.DisableDownload {
-		return http.StatusForbidden, fmt.Errorf("downloads are not allowed for this share")
-	}
-
-	// Check global download limit (if not using per-user limits)
-	if !d.Share.PerUserDownloadLimit && d.Share.DownloadsLimit > 0 && d.Share.Downloads >= d.Share.DownloadsLimit {
-		return http.StatusForbidden, fmt.Errorf("share downloads limit reached")
-	}
-
-	// Check per-user download limit
-	if d.Share.PerUserDownloadLimit {
-		// Block anonymous users
-		if d.User.Username == "anonymous" {
-			return http.StatusForbidden, fmt.Errorf("anonymous downloads are not allowed with per-user limits")
-		}
-		// Check if user has reached their limit
-		if d.Share.HasReachedUserLimit(d.User.Username) {
-			return http.StatusForbidden, fmt.Errorf("user download limit reached for this share")
-		}
-	}
-
-	if err := state.RecordShareDownload(d.Share.Hash, d.User.Username); err != nil {
-		return http.StatusInternalServerError, err
-	}
-
-	// Get all "file" parameter values (supports repeated params)
-	files := r.URL.Query()["file"]
-	if len(files) == 0 {
-		files = []string{"/"}
-	}
-
-	// Get the actual source name from the share's source mapping
-	sourceInfo, ok := settings.Config.Server.SourceMap[d.Share.SourcePath]
-	if !ok {
-		return http.StatusInternalServerError, fmt.Errorf("source not found for share")
-	}
-	actualSourceName := sourceInfo.Name
-
-	// Process each file path
-	fileList := []string{}
-	for _, file := range files {
-		// Rule 1: Validate each file path to prevent path traversal
-		cleanFile, err := utils.SanitizePath(file)
-		if err != nil {
-			return http.StatusBadRequest, fmt.Errorf("invalid file path: %v", err)
-		}
-
-		// Join the share path with the requested path
-		filePath := utils.JoinPathAsUnix(d.Share.Path, cleanFile)
-		fileList = append(fileList, filePath)
-	}
-
-	status, err := RawFilesHandler(w, r, d, actualSourceName, fileList)
-	if err != nil {
-		if err == errors.ErrDownloadNotAllowed {
-			return http.StatusForbidden, errors.ErrDownloadNotAllowed
-		}
-		logger.Errorf("public share handler: error processing filelist: %v with error %v", files, err)
-		return status, fmt.Errorf("error processing filelist: %v", files)
-	}
-	return status, nil
-}
-
-// publicShareHandler returns file or directory information from a public share.
-// @Summary Get file/directory information from a public share
-// @Description Returns metadata for files or directories accessible via a public share link. Browsing is disabled for upload-only shares.
-// @Tags Resources
-// @Accept json
-// @Produce json
-// @Param hash query string true "Share hash for authentication"
-// @Param path query string false "Path within the share to retrieve information for. Defaults to share root."
-// @Param content query string false "Include file content if true"
-// @Param metadata query string false "Extract audio/video metadata if true"
-// @Success 200 {object} iteminfo.FileInfo "File or directory metadata"
-// @Failure 403 {object} map[string]string "Share unavailable or access denied"
-// @Failure 404 {object} map[string]string "Share not found or file not found"
-// @Failure 500 {object} map[string]string "Internal server error"
-// @Failure 501 {object} map[string]string "Browsing disabled for upload shares"
-// @Router /public/api/resources [get]
-func publicGetResourceHandler(w http.ResponseWriter, r *http.Request, d *requestContext) (int, error) {
-	if d.Share.ShareType == "upload" {
-		return http.StatusNotImplemented, fmt.Errorf("browsing is disabled for upload shares")
-	}
-	return RenderJSON(w, r, d.FileInfo)
-}
-
-// publicUploadHandler processes file uploads to a public upload share.
-// @Summary Upload files to a public upload share
-// @Description Handles file and directory uploads to an upload-only public share. Supports chunked uploads, conflict resolution (override), and directory creation.
-// @Tags Resources
-// @Accept multipart/form-data
-// @Produce json
-// @Param hash query string true "Share hash for authentication"
-// @Param path query string true "path within the share to upload to. Must be relative to share root."
-// @Param override query bool false "If true, overwrite existing files/folders. Defaults to false."
-// @Param action query string false "Upload action: 'override' to replace files, 'rename' to auto-rename"
-// @Param file formData file true "File to upload"
-// @Success 200 {object} map[string]string "Upload successful"
-// @Failure 400 {object} map[string]string "Invalid request or parameters"
-// @Failure 403 {object} map[string]string "Share unavailable or upload not allowed"
-// @Failure 404 {object} map[string]string "Share not found"
-// @Failure 409 {object} map[string]string "File or directory already exists (conflict)"
-// @Failure 500 {object} map[string]string "Internal server error during upload"
-// @Failure 501 {object} map[string]string "Uploading disabled for non-upload shares"
-// @Router /public/api/resources [post]
-func publicUploadHandler(w http.ResponseWriter, r *http.Request, d *requestContext) (int, error) {
-	if d.Share.ShareType != "upload" && !d.Share.AllowCreate {
-		return http.StatusForbidden, fmt.Errorf("uploading is disabled for this share")
-	}
-	if !d.Share.AllowReplacements && r.URL.Query().Get("action") == "override" {
-		return http.StatusForbidden, fmt.Errorf("cannot overwrite files for this share")
-	}
-	// Go automatically decodes query params
-	source := settings.Config.Server.SourceMap[d.Share.SourcePath].Name
-	// adjust query params to match ResourcePostHandler
-	q := r.URL.Query()
-	q.Set("source", source)
-	q.Set("path", d.IndexPath)
-	r.URL.RawQuery = q.Encode()
-	status, err := ResourcePostHandler(w, r, d)
-	if err != nil {
-		logger.Errorf("public upload handler: error uploading with error %v", err)
-		return http.StatusInternalServerError, fmt.Errorf("upload failure occured on backend")
-	}
-	return status, nil
-}
-
-// health godoc
-// @Summary Health Check
-// @Schemes
-// @Description Returns the health status of the API.
-// @Accept json
-// @Produce json
-// @Success 200 {object} HttpResponse "successful health check response"
-// @Router /health [get]
-func healthHandler(w http.ResponseWriter, r *http.Request) {
-	w.Header().Set("Content-Type", "application/json")
-	response := HttpResponse{Message: "ok"}    // Create response with status "ok"
-	err := json.NewEncoder(w).Encode(response) // Encode the response into JSON
-	if err != nil {
-		http.Error(w, http.StatusText(http.StatusInternalServerError), http.StatusInternalServerError)
-	}
-}
-
-// publicPreviewHandler handles the preview request for images from public shares.
-// @Summary Get image/video preview from a public share
-// @Description Returns a preview (thumbnail) for images or videos accessible via a public share. Preview generation can be disabled globally or per-share. Not available for upload-only shares.
-// @Tags Resources
-// @Accept json
-// @Produce image/jpeg
-// @Param hash query string true "Share hash for authentication"
-// @Param path query string true "File path within the share to preview"
-// @Param size query string false "Preview size: 'small' or 'large'. Default is based on server settings.Config."
-// @Success 200 {file} file "Preview image content (JPEG)"
-// @Failure 403 {object} map[string]string "Share unavailable or access denied"
-// @Failure 404 {object} map[string]string "File not found or preview not available"
-// @Failure 500 {object} map[string]string "Internal server error"
-// @Failure 501 {object} map[string]string "Previews disabled globally, for this share, or for upload shares"
-// @Router /public/api/resources/preview [get]
-func publicPreviewHandler(w http.ResponseWriter, r *http.Request, d *requestContext) (int, error) {
-	if settings.Config.Server.DisablePreviews || d.Share.DisableThumbnails {
-		return http.StatusNotImplemented, fmt.Errorf("preview is disabled")
-	}
-	if d.Share.ShareType == "upload" {
-		return http.StatusNotImplemented, fmt.Errorf("preview is disabled for upload shares")
-	}
-	status, err := PreviewHelperFunc(w, r, d)
-	if err != nil {
-		logger.Errorf("public preview handler: error getting preview with error %v", err)
-		// Obfuscate errors for shares to prevent information leakage
-		return http.StatusNotFound, fmt.Errorf("preview not available for this item")
-	}
-	return status, err
-}
-
-// publicPutHandler handles the PUT request for a public share.
-// @Summary Update a file in a public share
-// @Description Updates the content of a file in a public share.
-// @Tags Resources
-// @Accept json
-// @Produce json
-// @Param hash query string true "Share hash for authentication"
-// @Param path query string true "Path to the file to update"
-// @Param content body string true "New content for the file"
-// @Success 200 {object} map[string]string "File updated successfully"
-// @Failure 400 {object} map[string]string "Invalid request or parameters"
-// @Failure 403 {object} map[string]string "Share unavailable or update not allowed"
-// @Failure 404 {object} map[string]string "Share not found or file not found"
-// @Failure 500 {object} map[string]string "Internal server error"
-// @Router /public/api/resources [put]
-func publicPutHandler(w http.ResponseWriter, r *http.Request, d *requestContext) (int, error) {
-	// update path to be the source path
-	if !d.Share.AllowModify {
-		return http.StatusForbidden, fmt.Errorf("create is not allowed for this share")
-	}
-	sourceName := d.Share.GetSourceName()
-	if sourceName == "" {
-		return http.StatusNotFound, fmt.Errorf("source not available")
-	}
-
-	if !d.Share.AllowModify {
-		return http.StatusForbidden, fmt.Errorf("edit permission not allowed for this share")
-	}
-	// Go automatically decodes query params
-	path := r.URL.Query().Get("path")
-
-	// Rule 1: Validate user-provided path to prevent path traversal
-	cleanPath, err := utils.SanitizePath(path)
-	if err != nil {
-		return http.StatusBadRequest, err
-	}
-
-	resolvedPath := utils.JoinPathAsUnix(d.Share.Path, cleanPath)
-	err = files.WriteFile(sourceName, resolvedPath, r.Body)
-	// hide the error
-	if err != nil {
-		logger.Errorf("public put handler: error updating resource with error %v", err)
-		return http.StatusInternalServerError, fmt.Errorf("an error occurred while updating the resource")
-	}
-	return http.StatusOK, nil
-}
-
-// deprecated -- see publicBulkDeleteHandler
-func publicDeleteHandler(w http.ResponseWriter, r *http.Request, d *requestContext) (int, error) {
-	if !d.Share.AllowDelete {
-		return http.StatusForbidden, fmt.Errorf("delete is not allowed for this share")
-	}
-	fileInfo, err := files.FileInfoFaster(utils.FileOptions{
-		FollowSymlinks: true,
-		Path:           d.IndexPath,
-		Source:         d.Share.SourcePath,
-	}, d.ShareUser)
-	if err != nil {
-		return http.StatusNotFound, fmt.Errorf("resource not available")
-	}
-	err = files.DeleteFiles(d.Share.SourcePath, fileInfo.RealPath, fileInfo.Type == "directory")
-	if err != nil {
-		logger.Errorf("public delete handler: error deleting resource with error %v", err)
-		return http.StatusInternalServerError, fmt.Errorf("an error occured while deleting the resource")
-	}
-	// delete thumbnails
-	preview.DelThumbs(r.Context(), *fileInfo)
-	return http.StatusOK, nil
-}
-
-// publicBulkDeleteHandler deletes multiple resources from a public share in a single request.
-// @Summary Bulk delete resources from public share
-// @Description Deletes multiple resources specified in the request body. Returns a list of succeeded and failed deletions.
-// @Tags Resources
-// @Accept json
-// @Produce json
-// @Param hash query string true "Share hash for authentication"
-// @Param items body []BulkDeleteItem true "Array of items to delete, each with source and path"
-// @Success 200 {object} BulkDeleteResponse "All resources deleted successfully"
-// @Success 207 {object} BulkDeleteResponse "Partial success - some resources deleted, some failed"
-// @Failure 400 {object} map[string]string "Bad request - invalid JSON or empty items array"
-// @Failure 403 {object} map[string]string "Forbidden - delete not allowed for this share"
-// @Failure 500 {object} map[string]string "Internal server error - all deletions failed"
-// @Router /public/api/resources/bulk [delete]
-func publicBulkDeleteHandler(w http.ResponseWriter, r *http.Request, d *requestContext) (int, error) {
-	if !d.Share.AllowDelete {
-		return http.StatusForbidden, fmt.Errorf("delete is not allowed for this share")
-	}
-	// hide the error
-	status, err := ResourceBulkDeleteHandler(w, r, d)
-	if err != nil {
-		logger.Errorf("public bulk delete handler: error deleting resources with error %v", err)
-		return http.StatusInternalServerError, fmt.Errorf("an error occurred while processing the request")
-	}
-	return status, nil
-}
-
-// publicPatchHandler performs a patch operation (e.g., move, copy, rename) on resources in a public share.
-// @Summary Move, copy, or rename resources in a public share
-// @Description Performs move, copy, or rename operations on multiple resources within a public share. All operations are performed atomically.
-// @Tags Resources
-// @Accept json
-// @Produce json
-// @Param hash query string true "Share hash for authentication"
-// @Param request body MoveCopyRequest true "Move/copy request with items and action"
-// @Success 200 {object} MoveCopyResponse "All operations completed successfully"
-// @Success 207 {object} MoveCopyResponse "Partial success - some operations succeeded, some failed"
-// @Failure 400 {object} map[string]string "Bad request - invalid JSON or parameters"
-// @Failure 403 {object} map[string]string "Forbidden - modify not allowed for this share"
-// @Failure 404 {object} map[string]string "Share or resource not found"
-// @Failure 500 {object} MoveCopyResponse "Internal server error"
-// @Router /public/api/resources [patch]
-func publicPatchHandler(w http.ResponseWriter, r *http.Request, d *requestContext) (int, error) {
-	if !d.Share.AllowModify {
-		return http.StatusForbidden, fmt.Errorf("edit permission not allowed for this share")
-	}
-
-	// Get the source from the share
-	sourceName := d.Share.GetSourceName()
-	if sourceName == "" {
-		return http.StatusNotFound, fmt.Errorf("source not available")
-	}
-
-	// Replace user with the share creator's user for proper permission checking
-	var shareCreatedByUser *users.User
-	userValue, err := state.UserForShareOwner(d.Share)
-	if err == nil {
-		shareCreatedByUser = &userValue
-	}
-	if err != nil {
-		return http.StatusNotFound, fmt.Errorf("user for share no longer exists")
-	}
-	d.User = shareCreatedByUser
-
-	// Parse the request body
-	var req MoveCopyRequest
-	if err = json.NewDecoder(r.Body).Decode(&req); err != nil {
-		return http.StatusBadRequest, fmt.Errorf("invalid JSON body: %v", err)
-	}
-
-	if req.Action == "" {
-		return http.StatusBadRequest, fmt.Errorf("action is required (copy, move, or rename)")
-	}
-
-	// Transform the request: prepend share path and add source to each item
-	// This normalizes the request to look like a regular user request
-	// Note: Share paths are absolute, so we don't strip user scope here
-	// resourcePatchHandler will skip adding scope for shares
-	for i := range req.Items {
-		var sanitizedPath string
-		sanitizedPath, err = utils.SanitizePath(req.Items[i].FromPath)
-		if err != nil {
-			return http.StatusBadRequest, fmt.Errorf("invalid from path: %w", err)
-		}
-		req.Items[i].FromSource = sourceName
-		req.Items[i].FromPath = utils.JoinPathAsUnix(d.Share.Path, sanitizedPath)
-		sanitizedPath, err = utils.SanitizePath(req.Items[i].ToPath)
-		if err != nil {
-			return http.StatusBadRequest, fmt.Errorf("invalid to path: %w", err)
-		}
-		req.Items[i].ToSource = sourceName
-		req.Items[i].ToPath = utils.JoinPathAsUnix(d.Share.Path, sanitizedPath)
-	}
-	d.Data = req
-
-	// Call the regular handler (will treat this like a normal user request now)
-	status, err := ResourcePatchHandler(w, r, d)
-
-	// For shares, we need to sanitize the response to hide internal details
-	// The response has already been written by resourcePatchHandler, but we can still return error
-	if err != nil {
-		logger.Errorf("public patch handler: error processing patch with error %v", err)
-		// Obfuscate errors for security
-		return http.StatusInternalServerError, fmt.Errorf("an error occurred while processing the request")
-	}
-
-	return status, err
-}
-
-// getShareImage serves banner or favicon files for shares as resizable previews
-// @Summary Get share image (banner or favicon) as preview
-// @Description Returns a resizable preview (large size) for the banner or favicon file of a share
-// @Tags Shares
-// @Produce image/jpeg
-// @Param hash query string true "Share hash"
-// @Param banner query bool false "Request banner file"
-// @Param favicon query bool false "Request favicon file"
-// @Success 200 {file} file "Preview image content (JPEG)"
-// @Failure 400 {object} map[string]string "Invalid request"
-// @Failure 403 {object} map[string]string "Permission denied"
-// @Failure 404 {object} map[string]string "Asset not found"
-// @Router /public/api/share/image [get]
-func getShareImage(w http.ResponseWriter, r *http.Request, d *requestContext) (int, error) {
-	// Determine which asset is being requested
-	isBanner := r.URL.Query().Get("banner") == "true"
-	isFavicon := r.URL.Query().Get("favicon") == "true"
-
-	if !isBanner && !isFavicon {
-		return http.StatusBadRequest, fmt.Errorf("either banner or favicon parameter must be true")
-	}
-
-	userValue, err := state.UserForShareOwner(d.Share)
-	var shareCreatedByUser *users.User
-	if err == nil {
-		shareCreatedByUser = &userValue
-	}
-	if err != nil {
-		return http.StatusNotFound, fmt.Errorf("user for share no longer exists")
-	}
-
-	sourceName, assetPath, err := d.Share.GetShareImagePartsHelper(isBanner)
-	if err != nil {
-		return http.StatusBadRequest, fmt.Errorf("invalid asset configuration: %v", err)
-	}
-
-	// Get file info
-	fileInfo, err := files.FileInfoFaster(utils.FileOptions{
-		Path:           assetPath,
-		Source:         sourceName,
-		Expand:         false,
-		Content:        false,
-		Metadata:       false,
-		ShowHidden:     false,
-		FollowSymlinks: true,
-	}, shareCreatedByUser)
-
-	if err != nil {
-		logger.Errorf("error accessing share asset: source=%v path=%v error=%v", sourceName, assetPath, err)
-		return http.StatusNotFound, fmt.Errorf("asset file not found or not accessible")
-	}
-
-	// Ensure it's an image file
-	if !strings.HasPrefix(fileInfo.Type, "image/") {
-		return http.StatusBadRequest, fmt.Errorf("invalid file type, must be image")
-	}
-
-	// Set file info in request context for preview generation
-	d.FileInfo = *fileInfo
-	q := r.URL.Query()
-	if isBanner {
-		q.Set("size", "xlarge")
-	} else {
-		q.Set("size", "small")
-	}
-	r.URL.RawQuery = q.Encode()
-
-	// Use the preview helper to generate and serve a resized preview
-	status, err := PreviewHelperFunc(w, r, d)
-	if err != nil {
-		logger.Errorf("error generating preview for share asset: source=%v path=%v error=%v", sourceName, assetPath, err)
-		return http.StatusNotFound, fmt.Errorf("preview not available for this asset")
-	}
-
-	return status, err
-}
-
-// publicItemsGetHandler efficiently returns a basic list of items for a directory in a public share.
-// @Summary Get directory items (public share)
-// @Description Efficiently returns a basic list of items for the specified path in a public share. Use hash for authentication instead of source. Use 'only' parameter to filter by only files or folders.
-// @Tags Resources
-// @Accept json
-// @Produce json
-// @Param hash query string true "Share hash for authentication"
-// @Param path query string false "Path within the share to list child items. Defaults to share root."
-// @Param only query string false "Filter: 'files', 'folders', or omit for both"
-// @Success 200 {object} files.Items "lists files and folders"
-// @Failure 403 {object} map[string]string "Forbidden (access denied)"
-// @Failure 404 {object} map[string]string "Share not found or source not found"
-// @Failure 500 {object} map[string]string "Internal server error"
-// @Router /public/api/resources/items [get]
-func publicItemsGetHandler(w http.ResponseWriter, r *http.Request, d *requestContext) (int, error) {
-	sourceInfo, ok := settings.Config.Server.SourceMap[d.Share.SourcePath]
-	if !ok {
-		return http.StatusNotFound, fmt.Errorf("source not found")
-	}
-	items, err := files.GetDirItems(utils.FileOptions{
-		FollowSymlinks: true,
-		Path:           d.IndexPath,
-		Source:         sourceInfo.Name,
-		ShowHidden:     d.ShareUser.ShowHidden,
-		Only:           r.URL.Query().Get("only"),
-	}, d.ShareUser)
-	if err != nil {
-		if err == errors.ErrAccessDenied {
-			return http.StatusForbidden, err
-		}
-		return http.StatusInternalServerError, err
-	}
-	return RenderJSON(w, r, items)
-}
