+++ backend/internal/web/stream.go
@@ -1,28 +1,268 @@
 package web
 
 import (
+	"errors"
 	"fmt"
 	"io"
+	"mime"
 	"net/http"
 	"os"
 	"path/filepath"
+	"strconv"
 	"strings"
+	"time"
 
+	"github.com/gtsteffaniak/filebrowser/backend/internal/state"
 	"github.com/gtsteffaniak/filebrowser/backend/internal/utils"
 	"github.com/gtsteffaniak/filebrowser/backend/pkg/indexing"
+	"github.com/gtsteffaniak/filebrowser/backend/pkg/indexing/iteminfo"
+	"github.com/gtsteffaniak/filebrowser/backend/pkg/settings"
 	"github.com/gtsteffaniak/go-logger/logger"
 	"golang.org/x/time/rate"
-	"github.com/gtsteffaniak/filebrowser/backend/pkg/settings"
-	"github.com/gtsteffaniak/filebrowser/backend/internal/state"
-
 )
+
+const (
+	viewGrantTTL        = 15 * time.Minute
+	maxStreamRangeBytes = 4 << 20 // 4 MiB
+)
+
+var errStreamRangeInvalid = errors.New("invalid byte range")
 
 type ServeSingleFileOptions struct {
 	ForceInline bool
 	RangeOnly   bool
 }
 
-// serveSingleFile opens one file and streams it with Range support via http.ServeContent.
+func normalizeViewGrantPath(p string) string {
+	return filepath.ToSlash(strings.TrimSpace(p))
+}
+
+func mintViewGrant(d *Context, source, filePath string) (string, error) {
+	token, err := utils.RandomHex(16)
+	if err != nil {
+		return "", err
+	}
+	grant := utils.ViewGrant{
+		UserID:    d.User.ID,
+		ShareHash: d.Share.Hash,
+		Source:    source,
+		Path:      normalizeViewGrantPath(filePath),
+		ExpiresAt: time.Now().Add(viewGrantTTL).Unix(),
+	}
+	utils.ViewGrantsCache.Set(token, grant)
+	return token, nil
+}
+
+func ValidateViewGrant(token string, d *Context, source, filePath string) error {
+	grant, ok := utils.ViewGrantsCache.Get(token)
+	if !ok {
+		return fmt.Errorf("invalid or expired view token")
+	}
+	if time.Now().Unix() > grant.ExpiresAt {
+		utils.ViewGrantsCache.Delete(token)
+		return fmt.Errorf("view token expired")
+	}
+	if grant.UserID != d.User.ID {
+		return fmt.Errorf("view token viewer mismatch")
+	}
+	if grant.ShareHash != d.Share.Hash {
+		return fmt.Errorf("view token share mismatch")
+	}
+	if grant.Source != source {
+		return fmt.Errorf("view token source mismatch")
+	}
+	if grant.Path != normalizeViewGrantPath(filePath) {
+		return fmt.Errorf("view token path mismatch")
+	}
+	perms, err := effectiveFilePerms(d, source)
+	if err != nil || !perms.View {
+		return fmt.Errorf("view permission required")
+	}
+	return nil
+}
+
+func canMintViewToken(d *Context, source string) bool {
+	perms, err := effectiveFilePerms(d, source)
+	return err == nil && perms.View
+}
+
+func AttachViewToken(d *Context, source, filePath string, file *iteminfo.ExtendedFileInfo) {
+	if file == nil || file.Type == "directory" {
+		return
+	}
+	if !canMintViewToken(d, source) {
+		return
+	}
+	token, err := mintViewGrant(d, source, filePath)
+	if err != nil {
+		return
+	}
+	file.ViewToken = token
+}
+
+func indexFilePath(dirPath, name string) string {
+	dirPath = normalizeViewGrantPath(dirPath)
+	if dirPath == "" || dirPath == "/" {
+		return "/" + name
+	}
+	if !strings.HasSuffix(dirPath, "/") {
+		dirPath += "/"
+	}
+	return dirPath + name
+}
+
+func AttachViewTokensForDirectory(d *Context, source, dirPath string, file *iteminfo.ExtendedFileInfo) {
+	if file == nil || file.Type != "directory" {
+		return
+	}
+	if !canMintViewToken(d, source) {
+		return
+	}
+	for i := range file.Files {
+		if file.Files[i].Type == "directory" {
+			continue
+		}
+		childPath := indexFilePath(dirPath, file.Files[i].Name)
+		token, err := mintViewGrant(d, source, childPath)
+		if err != nil {
+			continue
+		}
+		file.Files[i].ViewToken = token
+	}
+}
+
+// streamUseRangeOnly reports whether the stream endpoint must serve capped partial
+// content only (never a full-file 200 response). The stream endpoint is media-only.
+func streamUseRangeOnly(_ *Context, _ string) bool {
+	return true
+}
+
+func IsMediaStreamFile(displayFileName string) bool {
+	contentType := mime.TypeByExtension(strings.ToLower(filepathExt(displayFileName)))
+	return strings.HasPrefix(contentType, "video/") || strings.HasPrefix(contentType, "audio/")
+}
+
+func filepathExt(name string) string {
+	if i := strings.LastIndex(name, "."); i >= 0 {
+		return name[i:]
+	}
+	return ""
+}
+
+func parseStreamByteRange(rangeHeader string, size int64) (start, end int64, err error) {
+	if size <= 0 {
+		return 0, 0, errStreamRangeInvalid
+	}
+	if !strings.HasPrefix(rangeHeader, "bytes=") {
+		return 0, 0, errStreamRangeInvalid
+	}
+	spec := strings.TrimPrefix(rangeHeader, "bytes=")
+	if spec == "" || strings.Contains(spec, ",") {
+		return 0, 0, errStreamRangeInvalid
+	}
+
+	dash := strings.Index(spec, "-")
+	if dash < 0 {
+		return 0, 0, errStreamRangeInvalid
+	}
+	startStr := strings.TrimSpace(spec[:dash])
+	endStr := strings.TrimSpace(spec[dash+1:])
+
+	if startStr == "" {
+		// suffix range: bytes=-500
+		suffix, parseErr := strconv.ParseInt(endStr, 10, 64)
+		if parseErr != nil || suffix <= 0 {
+			return 0, 0, errStreamRangeInvalid
+		}
+		if suffix > size {
+			suffix = size
+		}
+		start = size - suffix
+		end = size - 1
+		return start, end, nil
+	}
+
+	start, err = strconv.ParseInt(startStr, 10, 64)
+	if err != nil || start < 0 || start >= size {
+		return 0, 0, errStreamRangeInvalid
+	}
+
+	if endStr == "" {
+		end = size - 1
+	} else {
+		end, err = strconv.ParseInt(endStr, 10, 64)
+		if err != nil || end < start {
+			return 0, 0, errStreamRangeInvalid
+		}
+		if end >= size {
+			end = size - 1
+		}
+	}
+	return start, end, nil
+}
+
+func capStreamByteRange(start, end int64) (int64, int64) {
+	if end-start+1 <= maxStreamRangeBytes {
+		return start, end
+	}
+	return start, start + maxStreamRangeBytes - 1
+}
+
+func setStreamResponseHeaders(w http.ResponseWriter, r *http.Request, displayFileName string, size int64) {
+	SetContentDisposition(w, r, displayFileName, true)
+	w.Header().Set("Accept-Ranges", "bytes")
+	w.Header().Set("Cache-Control", "private")
+	w.Header().Set("X-Content-Type-Options", "nosniff")
+	if contentType := mime.TypeByExtension(strings.ToLower(filepathExt(displayFileName))); contentType != "" {
+		w.Header().Set("Content-Type", contentType)
+	}
+	w.Header().Set("Content-Length", strconv.FormatInt(size, 10))
+}
+
+func serveStreamByteRange(w http.ResponseWriter, r *http.Request, reader io.ReadSeeker, displayFileName string, size int64) (int, error) {
+	if r.Method == http.MethodHead {
+		setStreamResponseHeaders(w, r, displayFileName, size)
+		return http.StatusOK, nil
+	}
+
+	rangeHeader := r.Header.Get("Range")
+	if rangeHeader == "" {
+		setStreamResponseHeaders(w, r, displayFileName, size)
+		w.Header().Set("Content-Range", fmt.Sprintf("bytes */%d", size))
+		return http.StatusRequestedRangeNotSatisfiable, fmt.Errorf("stream requires byte range requests")
+	}
+
+	start, end, err := parseStreamByteRange(rangeHeader, size)
+	if err != nil {
+		setStreamResponseHeaders(w, r, displayFileName, size)
+		w.Header().Set("Content-Range", fmt.Sprintf("bytes */%d", size))
+		return http.StatusRequestedRangeNotSatisfiable, fmt.Errorf("invalid byte range")
+	}
+	start, end = capStreamByteRange(start, end)
+
+	chunkSize := end - start + 1
+	if _, err := reader.Seek(start, io.SeekStart); err != nil {
+		return http.StatusInternalServerError, err
+	}
+
+	SetContentDisposition(w, r, displayFileName, true)
+	w.Header().Set("Accept-Ranges", "bytes")
+	w.Header().Set("Cache-Control", "private")
+	w.Header().Set("X-Content-Type-Options", "nosniff")
+	if contentType := mime.TypeByExtension(strings.ToLower(filepathExt(displayFileName))); contentType != "" {
+		w.Header().Set("Content-Type", contentType)
+	}
+	w.Header().Set("Content-Range", fmt.Sprintf("bytes %d-%d/%d", start, end, size))
+	w.Header().Set("Content-Length", strconv.FormatInt(chunkSize, 10))
+	w.WriteHeader(http.StatusPartialContent)
+
+	if _, err := io.CopyN(w, reader, chunkSize); err != nil && !errors.Is(err, io.EOF) {
+		return http.StatusPartialContent, err
+	}
+	return http.StatusPartialContent, nil
+}
+
+// ServeSingleFile opens one file and streams it with Range support via http.ServeContent.
 func ServeSingleFile(w http.ResponseWriter, r *http.Request, d *Context, source string, scopedFilePath string, displayFileName string, opts ServeSingleFileOptions) (int, error) {
 	idx := indexing.GetIndex(source)
 	if idx == nil {
@@ -216,7 +456,7 @@ func publicStreamHandler(w http.ResponseWriter, r *http.Request, d *Context) (in
 	return status, nil
 }
 
-// resolveDisplayFileList returns client-facing paths for activity logging.
+// ResolveDisplayFileList returns client-facing paths for activity logging.
 func ResolveDisplayFileList(d *Context, source string, fileList []string) []string {
 	if d.Share.Hash != "" {
 		display := make([]string, 0, len(fileList))
