+++ backend/internal/web/stream_test.go
@@ -2,8 +2,11 @@ package web
 
 import (
 	"fmt"
+	"io"
 	"net/http"
 	"net/http/httptest"
+	"os"
+	"strings"
 	"sync"
 	"testing"
 	"time"
@@ -14,12 +17,12 @@ import (
 	"github.com/gtsteffaniak/filebrowser/backend/pkg/indexing/iteminfo"
 )
 
-var streamGrantResolverMu sync.Mutex
+var streamTestResolverMu sync.Mutex
 
-func initStreamGrantTestSources(t *testing.T) {
+func initStreamTestSources(t *testing.T) {
 	t.Helper()
-	streamGrantResolverMu.Lock()
-	t.Cleanup(func() { streamGrantResolverMu.Unlock() })
+	streamTestResolverMu.Lock()
+	t.Cleanup(func() { streamTestResolverMu.Unlock() })
 	users.SetSourceNameResolver(func(name string) (string, error) {
 		switch name {
 		case "default", "Downloads", "srv":
@@ -74,7 +77,7 @@ func testUserWithView(id uint64, sources ...string) *users.User {
 
 func TestMintAndValidateViewGrant(t *testing.T) {
 	t.Parallel()
-	initStreamGrantTestSources(t)
+	initStreamTestSources(t)
 	d := &requestContext{User: testUserWithView(42, "default")}
 	token, err := mintViewGrant(d, "default", "/docs/track.mp3")
 	if err != nil {
@@ -90,7 +93,7 @@ func TestMintAndValidateViewGrant(t *testing.T) {
 
 func TestValidateViewGrantWrongUser(t *testing.T) {
 	t.Parallel()
-	initStreamGrantTestSources(t)
+	initStreamTestSources(t)
 	owner := &requestContext{User: testUserWithView(1, "default")}
 	other := &requestContext{User: testUserWithView(2, "default")}
 	token, err := mintViewGrant(owner, "default", "/a.mp3")
@@ -104,7 +107,7 @@ func TestValidateViewGrantWrongUser(t *testing.T) {
 
 func TestValidateViewGrantWrongPath(t *testing.T) {
 	t.Parallel()
-	initStreamGrantTestSources(t)
+	initStreamTestSources(t)
 	d := &requestContext{User: testUserWithView(1, "default")}
 	token, err := mintViewGrant(d, "default", "/a.mp3")
 	if err != nil {
@@ -117,7 +120,7 @@ func TestValidateViewGrantWrongPath(t *testing.T) {
 
 func TestValidateViewGrantExpired(t *testing.T) {
 	t.Parallel()
-	initStreamGrantTestSources(t)
+	initStreamTestSources(t)
 	token, err := utils.RandomHex(16)
 	if err != nil {
 		t.Fatal(err)
@@ -136,7 +139,7 @@ func TestValidateViewGrantExpired(t *testing.T) {
 
 func TestValidateViewGrantShareBinding(t *testing.T) {
 	t.Parallel()
-	initStreamGrantTestSources(t)
+	initStreamTestSources(t)
 	d := &requestContext{
 		User:  testUserWithView(1, "srv"),
 		Share: share.Share{ShareColumns: share.ShareColumns{Hash: "abc123"}},
@@ -156,7 +159,7 @@ func TestValidateViewGrantShareBinding(t *testing.T) {
 
 func TestAttachViewTokenForAllFileTypes(t *testing.T) {
 	t.Parallel()
-	initStreamGrantTestSources(t)
+	initStreamTestSources(t)
 	d := &requestContext{User: testUserWithView(7, "default")}
 	audio := &iteminfo.ExtendedFileInfo{
 		FileInfo: iteminfo.FileInfo{
@@ -180,7 +183,7 @@ func TestAttachViewTokenForAllFileTypes(t *testing.T) {
 
 func TestAttachViewTokensForDirectory(t *testing.T) {
 	t.Parallel()
-	initStreamGrantTestSources(t)
+	initStreamTestSources(t)
 	d := &requestContext{User: testUserWithView(7, "Downloads")}
 	file := &iteminfo.ExtendedFileInfo{
 		FileInfo: iteminfo.FileInfo{
@@ -212,7 +215,7 @@ func TestAttachViewTokensForDirectory(t *testing.T) {
 
 func TestAttachViewTokenDeniedWithoutViewPermission(t *testing.T) {
 	t.Parallel()
-	initStreamGrantTestSources(t)
+	initStreamTestSources(t)
 	d := &requestContext{
 		User: &users.User{
 			ID: 9,
@@ -235,7 +238,7 @@ func TestAttachViewTokenDeniedWithoutViewPermission(t *testing.T) {
 
 func TestStreamHandlerRejectsMissingToken(t *testing.T) {
 	t.Parallel()
-	initStreamGrantTestSources(t)
+	initStreamTestSources(t)
 	d := &requestContext{User: testUserWithView(1, "default")}
 	req := httptest.NewRequest(http.MethodGet, "/api/media/stream?source=default&file=/a.mp3", nil)
 	status, err := streamHandler(httptest.NewRecorder(), req, d)
@@ -246,7 +249,7 @@ func TestStreamHandlerRejectsMissingToken(t *testing.T) {
 
 func TestStreamHandlerRejectsNonMedia(t *testing.T) {
 	t.Parallel()
-	initStreamGrantTestSources(t)
+	initStreamTestSources(t)
 	d := &requestContext{User: testUserWithView(1, "default")}
 	token, err := mintViewGrant(d, "default", "/doc.pdf")
 	if err != nil {
@@ -261,7 +264,7 @@ func TestStreamHandlerRejectsNonMedia(t *testing.T) {
 
 func TestViewHandlerRejectsMedia(t *testing.T) {
 	t.Parallel()
-	initStreamGrantTestSources(t)
+	initStreamTestSources(t)
 	d := &requestContext{User: testUserWithView(1, "default")}
 	token, err := mintViewGrant(d, "default", "/clip.mp4")
 	if err != nil {
@@ -276,7 +279,7 @@ func TestViewHandlerRejectsMedia(t *testing.T) {
 
 func TestViewHandlerRejectsMissingToken(t *testing.T) {
 	t.Parallel()
-	initStreamGrantTestSources(t)
+	initStreamTestSources(t)
 	d := &requestContext{User: testUserWithView(1, "default")}
 	req := httptest.NewRequest(http.MethodGet, "/api/resources/view?source=default&file=/a.txt", nil)
 	status, err := viewHandler(httptest.NewRecorder(), req, d)
@@ -287,7 +290,7 @@ func TestViewHandlerRejectsMissingToken(t *testing.T) {
 
 func TestStreamHandlerRejectsMultiFile(t *testing.T) {
 	t.Parallel()
-	initStreamGrantTestSources(t)
+	initStreamTestSources(t)
 	d := &requestContext{User: testUserWithView(1, "default")}
 	req := httptest.NewRequest(http.MethodGet, "/api/media/stream?source=default&file=/a.mp3&file=/b.mp3&viewToken=tok", nil)
 	status, err := streamHandler(httptest.NewRecorder(), req, d)
@@ -298,7 +301,7 @@ func TestStreamHandlerRejectsMultiFile(t *testing.T) {
 
 func TestStreamHandlerRejectsArchiveParams(t *testing.T) {
 	t.Parallel()
-	initStreamGrantTestSources(t)
+	initStreamTestSources(t)
 	d := &requestContext{User: testUserWithView(1, "default")}
 	req := httptest.NewRequest(http.MethodGet, "/api/media/stream?source=default&file=/a.mp3&viewToken=tok&algo=zip", nil)
 	status, err := streamHandler(httptest.NewRecorder(), req, d)
@@ -314,11 +317,172 @@ func TestStreamHandlerRejectsArchiveParams(t *testing.T) {
 
 func TestViewHandlerRejectsArchiveParams(t *testing.T) {
 	t.Parallel()
-	initStreamGrantTestSources(t)
+	initStreamTestSources(t)
 	d := &requestContext{User: testUserWithView(1, "default")}
 	req := httptest.NewRequest(http.MethodGet, "/api/resources/view?source=default&file=/a.txt&viewToken=tok&algo=zip", nil)
 	status, err := viewHandler(httptest.NewRecorder(), req, d)
 	if status != http.StatusForbidden || err == nil {
 		t.Fatalf("expected 403 for algo param, got status=%d err=%v", status, err)
+	}
+}
+
+func TestParseStreamByteRange(t *testing.T) {
+	t.Parallel()
+	tests := []struct {
+		name      string
+		header    string
+		size      int64
+		wantStart int64
+		wantEnd   int64
+		wantErr   bool
+	}{
+		{name: "closed range", header: "bytes=0-99", size: 1000, wantStart: 0, wantEnd: 99},
+		{name: "open ended", header: "bytes=10-", size: 100, wantStart: 10, wantEnd: 99},
+		{name: "suffix", header: "bytes=-50", size: 100, wantStart: 50, wantEnd: 99},
+		{name: "missing prefix", header: "0-99", size: 100, wantErr: true},
+		{name: "multipart", header: "bytes=0-1,2-3", size: 100, wantErr: true},
+		{name: "start past end", header: "bytes=100-", size: 100, wantErr: true},
+	}
+	for _, tc := range tests {
+		tc := tc
+		t.Run(tc.name, func(t *testing.T) {
+			t.Parallel()
+			start, end, err := parseStreamByteRange(tc.header, tc.size)
+			if tc.wantErr {
+				if err == nil {
+					t.Fatal("expected error")
+				}
+				return
+			}
+			if err != nil {
+				t.Fatalf("parseStreamByteRange: %v", err)
+			}
+			if start != tc.wantStart || end != tc.wantEnd {
+				t.Fatalf("got %d-%d, want %d-%d", start, end, tc.wantStart, tc.wantEnd)
+			}
+		})
+	}
+}
+
+func TestCapStreamByteRange(t *testing.T) {
+	t.Parallel()
+	start, end := capStreamByteRange(0, maxStreamRangeBytes)
+	if end-start+1 != maxStreamRangeBytes {
+		t.Fatalf("expected capped range of %d bytes, got %d", maxStreamRangeBytes, end-start+1)
+	}
+	start, end = capStreamByteRange(100, 150)
+	if start != 100 || end != 150 {
+		t.Fatalf("unexpected cap for small range: %d-%d", start, end)
+	}
+}
+
+func TestStreamUseRangeOnly(t *testing.T) {
+	t.Parallel()
+	mediaCtx := &requestContext{User: &users.User{FrontendUser: users.FrontendUser{Permissions: users.Permissions{Download: true}}}}
+	if !streamUseRangeOnly(mediaCtx, "clip.mp4") {
+		t.Fatal("expected range-only for stream endpoint")
+	}
+	if !streamUseRangeOnly(mediaCtx, "notes.txt") {
+		t.Fatal("stream endpoint is always range-only")
+	}
+}
+
+func TestServeStreamByteRangeRejectsFullGET(t *testing.T) {
+	t.Parallel()
+	body := strings.Repeat("a", 256)
+	file, err := os.CreateTemp("", "stream-range-*")
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer os.Remove(file.Name())
+	if _, err = file.WriteString(body); err != nil {
+		t.Fatal(err)
+	}
+	if _, err = file.Seek(0, io.SeekStart); err != nil {
+		t.Fatal(err)
+	}
+
+	req := httptest.NewRequest(http.MethodGet, "/stream", nil)
+	rec := httptest.NewRecorder()
+	status, err := serveStreamByteRange(rec, req, file, "clip.mp4", int64(len(body)))
+	if status != http.StatusRequestedRangeNotSatisfiable || err == nil {
+		t.Fatalf("expected 416, got status=%d err=%v", status, err)
+	}
+}
+
+func TestServeStreamByteRangeReturnsPartialContent(t *testing.T) {
+	t.Parallel()
+	body := strings.Repeat("b", 512)
+	file, err := os.CreateTemp("", "stream-range-*")
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer os.Remove(file.Name())
+	if _, err = file.WriteString(body); err != nil {
+		t.Fatal(err)
+	}
+	if _, err = file.Seek(0, io.SeekStart); err != nil {
+		t.Fatal(err)
+	}
+
+	req := httptest.NewRequest(http.MethodGet, "/stream", nil)
+	req.Header.Set("Range", "bytes=0-99")
+	rec := httptest.NewRecorder()
+	status, err := serveStreamByteRange(rec, req, file, "clip.mp4", int64(len(body)))
+	if err != nil {
+		t.Fatalf("serveStreamByteRange: %v", err)
+	}
+	if status != http.StatusPartialContent {
+		t.Fatalf("expected 206, got %d", status)
+	}
+	if rec.Code != http.StatusPartialContent {
+		t.Fatalf("recorder code: %d", rec.Code)
+	}
+	if got := rec.Body.Len(); got != 100 {
+		t.Fatalf("expected 100 bytes body, got %d", got)
+	}
+	if !strings.HasPrefix(rec.Header().Get("Content-Range"), "bytes 0-99/512") {
+		t.Fatalf("unexpected Content-Range: %q", rec.Header().Get("Content-Range"))
+	}
+}
+
+func TestServeStreamByteRangeCapsOpenEndedRange(t *testing.T) {
+	t.Parallel()
+	size := int(maxStreamRangeBytes + 1024)
+	body := strings.Repeat("c", size)
+	file, err := os.CreateTemp("", "stream-range-*")
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer os.Remove(file.Name())
+	if _, err = file.WriteString(body); err != nil {
+		t.Fatal(err)
+	}
+	if _, err = file.Seek(0, io.SeekStart); err != nil {
+		t.Fatal(err)
+	}
+
+	req := httptest.NewRequest(http.MethodGet, "/stream", nil)
+	req.Header.Set("Range", "bytes=0-")
+	rec := httptest.NewRecorder()
+	status, err := serveStreamByteRange(rec, req, file, "song.mp3", int64(len(body)))
+	if err != nil {
+		t.Fatalf("serveStreamByteRange: %v", err)
+	}
+	if status != http.StatusPartialContent {
+		t.Fatalf("expected 206, got %d", status)
+	}
+	if rec.Body.Len() != int(maxStreamRangeBytes) {
+		t.Fatalf("expected capped body %d, got %d", maxStreamRangeBytes, rec.Body.Len())
+	}
+}
+
+func TestIsMediaStreamFile(t *testing.T) {
+	t.Parallel()
+	if !IsMediaStreamFile("movie.mp4") || !IsMediaStreamFile("track.flac") {
+		t.Fatal("expected media extensions to match")
+	}
+	if IsMediaStreamFile("readme.txt") {
+		t.Fatal("did not expect text file to match")
 	}
 }
