--- backend/pkg/indexing/hooks.go
@@ -1,4 +1,0 @@
-package indexing
-
-// OnSourceRootFullScanComplete is invoked after a source root ("/") full scan finishes.
-var OnSourceRootFullScanComplete func(sourceName string)
