+++ backend/pkg/indexing/indexingScanner.go
@@ -122,9 +122,6 @@ func (s *Scanner) runRootScan(quick bool) {
 		s.fullScanTime = int(time.Since(startTime).Seconds())
 		s.statsMu.Unlock()
 		s.updateComplexity()
-		if OnSourceRootFullScanComplete != nil {
-			OnSourceRootFullScanComplete(s.idx.Name)
-		}
 	}
 
 	s.statsMu.Lock()
